Refactor CML client into Trireme
