+++ sample-dApps/always-succeeds-contract/src/logic/script.rs
@@ -2,9 +2,9 @@ use cardano_multiplatform_lib::{
     address::{EnterpriseAddress, StakeCredential},
     plutus::{PlutusScript, PlutusV1Script},
 };
+use naumachia::trireme_ledger_client::cml_client::validator_script::PlutusScriptFile;
 use naumachia::{
     address::Address,
-    ledger_client::cml_client::validator_script::PlutusScriptFile,
     scripts::{ScriptError, ScriptResult, TxContext, ValidatorCode},
 };
