+++ src/ledger_client.rs
@@ -1,6 +1,5 @@
 use thiserror::Error;
 
-pub mod cml_client;
 pub mod in_memory_ledger;
 pub mod local_persisted_ledger;
 use async_trait::async_trait;
