+++ src/trireme_ledger_client/cml_client/tests.rs
@@ -1,11 +1,9 @@
 use super::*;
-use crate::{
-    ledger_client::cml_client::{
-        blockfrost_ledger::BlockFrostLedger,
-        key_manager::{KeyManager, TESTNET},
-    },
-    PolicyId,
+use crate::trireme_ledger_client::cml_client::{
+    blockfrost_ledger::BlockFrostLedger,
+    key_manager::{KeyManager, TESTNET},
 };
+use crate::PolicyId;
 use blockfrost_http_client::load_key_from_file;
 use cardano_multiplatform_lib::address::BaseAddress;
 use std::time::Duration;
